fix set_xgpu_dev.js in order to get a runnable project if you clone the repo

+++ set_xgpu_dev.js
@@ -7,16 +7,21 @@ const currentDir = path.dirname(fileURLToPath(import.meta.url));
 
 // Chemin vers le projet xgpu
 const XGPU_PATH = path.join(currentDir, '../xgpu');
-const PACKAGE_JSON_PATH = pathToFileURL(path.join(XGPU_PATH, 'package.json')).href;
 
-// Lire le package.json de xgpu de manière synchrone
-const packageJsonData = fs.readFileSync(fileURLToPath(PACKAGE_JSON_PATH), 'utf-8');
-const packageJson = JSON.parse(packageJsonData);
+if (fs.existsSync(XGPU_PATH) && fs.statSync(XGPU_PATH).isDirectory()) {
+
+    const PACKAGE_JSON_PATH = pathToFileURL(path.join(XGPU_PATH, 'package.json')).href;
+
+    // Lire le package.json de xgpu de manière synchrone
+    const packageJsonData = fs.readFileSync(fileURLToPath(PACKAGE_JSON_PATH), 'utf-8');
+    const packageJson = JSON.parse(packageJsonData);
 
 
-// Modifier les champs nécessaires
-packageJson.main = "src/index.ts";
-packageJson.types = "src/index.ts";
+    // Modifier les champs nécessaires
+    packageJson.main = "src/index.ts";
+    packageJson.types = "src/index.ts";
 
-// Écrire les modifications dans le package.json de xgpu
-fs.writeFileSync(fileURLToPath(PACKAGE_JSON_PATH), JSON.stringify(packageJson, null, 2));
+    // Écrire les modifications dans le package.json de xgpu
+    fs.writeFileSync(fileURLToPath(PACKAGE_JSON_PATH), JSON.stringify(packageJson, null, 2));
+
+}
